fix: extract add coupon form and coupon table to separate components
