+++ src/BeanVault.Client/src/types/Coupon.ts
@@ -1,0 +1,6 @@
+export type Coupon = {
+  id: string;
+  couponCode: string;
+  discountAmount: number;
+  minAmount: number;
+};
